backend/requirements.txt
# Core Flask dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Machine Learning dependencies
torch>=2.6.0
transformers>=4.40.0
huggingface-hub>=0.17.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Utility dependencies
python-dotenv>=1.0.0

# Optional: face-recognition>=1.3.0 (for face detection, requires CMake)